made utils folder and refactored
