+++ pages/api/admin/authentication/index.tsx
@@ -4,7 +4,7 @@ import { Filter, WithId } from "mongodb";
 import { Admin, ResponseMessage } from "../../../../models/types2";
 import { ADMIN_COLLECTION, INVITE_COLLECTION, LOGIN_IRON_OPTION } from "../../../../models/constants";
 import { withIronSessionApiRoute } from "iron-session/next";
-import { isInviteValid } from "../../../../db/utils";
+import { isInviteValid } from "../../../../util/utils";
 
 export default withIronSessionApiRoute(async function handler(
     req: NextApiRequest,
