+++ pages/api/admin/forgotPassword.ts
@@ -1,0 +1,8 @@
+import { NextApiHandler } from "next"
+
+const handler : NextApiHandler = (req, res) => {
+    const email: string = req.body?.email ?? "";
+    
+}
+
+export default handler
